Crate dupe

A cheap version of Clone.

Traits

Dupe

Like Clone, but should only be available if Clone is constant time and zero allocation (e.g. a few Arc bumps). The implementation of dupe should always call clone.

IterDupedExt

OptionDupedExt

Extension traits on Option where it holds a ref.

Derive Macros

Clone_

Derive the Clone trait, but without requiring all type arguments to implement [Clone].

Copy_

Derive the Copy trait, but without requiring all type arguments to implement [Copy].

Dupe

Derive the Dupe trait.

Dupe_

Derive the Dupe trait, but without requiring all type arguments to implement Dupe.