dupe

Trait IterDupedExt

pub trait IterDupedExt: Sized {
    // Required method
    fn duped(self) -> Cloned<Self>;
}

Required Methods

fn duped(self) -> Cloned<Self>

Like duped(), but only works for types that implement Dupe. Note that the return type is deliberately Cloned, as that behaves the same as a Duped would be, but can take advantage of standard library optimisations.

use std::rc::Rc;

use dupe::IterDupedExt;
let inputs = vec![Rc::new("Hello"), Rc::new("World")];
let outputs = inputs.iter().duped().collect::<Vec<_>>();
assert_eq!(inputs, outputs);

use gazebo::prelude::*; use std::cmp::Ordering;

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.